mat5.h: Changed return value of WriteCompressedStructField to size_t
mat5.c (GetStructFieldBufSize): Added

mat5.c (GetMatrixMaxBufSize): Changed function used to calculate the number of bytes in a struct field

mat5.c (WriteCompressedStructField): Updated MAT_C_STRUCT case for compressed output

mat5.c (Write5): Fixed tag in uncompressed output when the variable name is 4 chars or less

mat5.c (Write5): Updated MAT_C_STRUCT case for compressed output

test_mat.c (test_write_compressed_struct): Fixed the test and added help documentation for it

test_mat.c: Renamed writestruct test to write_struct

+++ src/mat5.h
@@ -29,7 +29,7 @@
 EXTERN int WriteCellArrayField( mat_t *mat, matvar_t *matvar,
                                 int compress );
 #if defined(HAVE_ZLIB)
-EXTERN int WriteCompressedStructField(mat_t *mat,matvar_t *matvar,z_stream *z);
+EXTERN size_t WriteCompressedStructField(mat_t *mat,matvar_t *matvar,z_stream *z);
 #endif
 EXTERN int WriteStructField(mat_t *mat,matvar_t *matvar);
 EXTERN int ReadNextStructField( mat_t *mat, matvar_t *matvar );
